+++ src/Model/QuantityBased.php
@@ -64,6 +64,7 @@ class QuantityBased extends PostageType
     {
         $return = ArrayList::create();
         $locations = $this->Locations();
+        $exclude = $this->Exclusions();
         $country = $parcel->getCountry();
         $region = $parcel->getRegion();
         $value = (float)$parcel->getItems();
@@ -73,17 +74,29 @@ class QuantityBased extends PostageType
         if ($this->Tax()->exists()) {
             $tax = $this->Tax()->ValidTax();
         }
-        
-        // Should this type filter based on location
-        if (!$locations->exists()) {
-            $check = true;
-        } elseif (isset($country) && isset($region)) {
-            $locations = $locations->filter([
-                "Regions.CountryCode" => $country,
-                "Regions.Code" => $region
-            ]);
 
+        if ($exclude->exists()) {
+            $exclude = $exclude->filter(
+                [
+                    "Regions.CountryCode" => $country,
+                    "Regions.Code" => $region
+                ]
+            );
+        }
+
+        if (!$exclude->exists()) {
             if ($locations->exists()) {
+                $locations = $locations->filter(
+                    [
+                        "Regions.CountryCode" => $country,
+                        "Regions.Code" => $region
+                    ]
+                );
+
+                if ($locations->exists()) {
+                    $check = true;
+                }
+            } else {
                 $check = true;
             }
         }