+++ changelog.md
@@ -17,3 +17,8 @@ Add `PostageForm` for scaffolding postage queries
 ## 1.1.1
 
 Fix error with weights/value
+
+## 1.2.0
+
+* Move locations to `PostageType` instead of individual postage types.
+* Add ability to "exclude" zones from a postage type (for example, all zones except the UK).